# docs/swagger.json
{
	"components": {
...
		"securitySchemes": {}
	},
	"info": {
		"title": "express_typescript_skeleton",
		"version": "1.0.0",
		"description": "a basic express app writen in typescript",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Abdulraheem A omeiza"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/user/pre-register": {
			"post": {
				"operationId": "PreRegisteruser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"emailOrMobile": "09053634327"
										}
									},
									"Example 2": {
										"value": {
											"emailOrMobile": "blesseth.omeiza@gmail.com",
											"unit": "usher"
										}
									}
								}
							}
						}
					},
					"201": {
						"description": "user pre-registered successfully"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/user/check-invitation": {
			"get": {
				"operationId": "CheckInvitation",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"413": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/user/register": {
			"post": {
				"operationId": "CreateLogin",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/user/verify-token": {
			"post": {
				"operationId": "VerifyAccount",
				"responses": {
					"200": {
						"description": "Account verified"
					},
					"201": {
						"description": "Account verified",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"token": 54875,
											"tokenRoute": "email"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/user/send-token": {
			"post": {
				"operationId": "SendToken",
				"responses": {
					"201": {
						"description": "Token sent",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"user": "+2347044124767",
											"tokenRoute": "mobile"
										}
									},
									"Example 2": {
										"value": {
											"user": "blesseth.omeiza@gmail.com",
											"tokenRoute": "email"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/user/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Login successfull",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"emailOrMobile": "blesseth.omeiza@gmail.com",
											"password": "123456qwerty"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														false
													],
													"nullable": false
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														false
													],
													"nullable": false
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														false
													],
													"nullable": false
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														false
													],
													"nullable": false
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}